rework _accel types
